stacking reward

+++ dreamerv2/train.py
@@ -109,7 +109,7 @@ def main():
     contact_force_sum = float(ep['log_contact_forces'].astype(np.float64).sum())
     contact_force_mean = float(ep['log_contact_forces'].astype(np.float64).mean())
     # Box Pos:
-    box_pos_z_mean = float(ep['log_box_pos_z_mean'].astype(np.uint32).mean())
+    box_pos_z_mean = float(ep['log_box_pos_z_mean'].astype(np.float64).mean())
 
     print(f'{mode.title()} episode has {length} steps and return {score:.1f}, grab reward {grab_reward:.1f} and stacking reward {stacking_reward:.1f}.')
     print(f'Episode has {contacts} contacts and contact force sum {contact_force_sum:.1f} and mean {contact_force_mean:.1f}.')
